Refactored new toasting framework into ActionDisplay
